requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
azure-ai-inference
azure-core
pydantic[email]
python-multipart
requests
